requirements.txt
djangorestframework-simplejwt==5.5.0
drf-nested-routers
